requirements.txt
django>=5.0
djangorestframework
django-cors-headers
openai
python-dotenv
freeflow-llm
